refactor: extract key/identity utilities from nostr-helper-functions
