+++ components/__tests__/display-products.test.tsx
@@ -635,8 +635,7 @@ describe("DisplayProducts search filtering", () => {
       fetch: jest.fn().mockResolvedValue([relayProduct]),
     };
     let setContextProductEvents:
-      | Dispatch<SetStateAction<NostrEvent[]>>
-      | undefined;
+      Dispatch<SetStateAction<NostrEvent[]>> | undefined;
 
     function StatefulDisplayProducts() {
       const [contextProductEvents, setProductEvents] = useState<NostrEvent[]>(
