+++ components/messages/__tests__/messages.test.tsx
@@ -10,6 +10,7 @@ import Messages from "../messages";
 import { ChatsContext } from "../../../utils/context/context";
 import { SignerContext } from "@/components/utility-components/nostr-context-provider";
 import * as nostrHelper from "@/utils/nostr/nostr-helper-functions";
+import * as keyUtils from "@/utils/nostr/key-utilities";
 import * as keypressHandler from "@/utils/keypress-handler";
 import { useRouter } from "next/router";
 
@@ -80,9 +81,11 @@ jest.mock("nostr-tools", () => ({
 }));
 
 jest.mock("@/utils/nostr/nostr-helper-functions");
+jest.mock("@/utils/nostr/key-utilities");
 jest.mock("@/utils/keypress-handler");
 
 const mockNostrHelper = nostrHelper as jest.Mocked<typeof nostrHelper>;
+const mockKeyUtils = keyUtils as jest.Mocked<typeof keyUtils>;
 const mockUseKeyPress = keypressHandler.useKeyPress as jest.Mock;
 
 describe("Messages Component", () => {
@@ -156,13 +159,11 @@ describe("Messages Component", () => {
     };
     (useRouter as jest.Mock).mockReturnValue(mockRouter);
 
-    mockNostrHelper.generateKeys.mockResolvedValue({
+    mockKeyUtils.generateKeys.mockResolvedValue({
       nsec: "testnsec",
       npub: "testnpub",
     });
-    mockNostrHelper.decryptNpub.mockImplementation(
-      (npub) => `${npub}-decrypted`
-    );
+    mockKeyUtils.decryptNpub.mockImplementation((npub) => `${npub}-decrypted`);
 
     mockUseKeyPress.mockReturnValue(false);
   });
@@ -240,7 +241,7 @@ describe("Messages Component", () => {
     mockRouter.query.pk = "broken_npub";
     mockChatsContextValue.isLoading = false;
     mockChatsContextValue.chatsMap = mockChatsMap;
-    mockNostrHelper.decryptNpub.mockReturnValue(null as any);
+    mockKeyUtils.decryptNpub.mockReturnValue(null as any);
 
     renderComponent();
