+++ components/product-invoice-card.tsx
@@ -49,6 +49,7 @@ import {
   isTimeoutError,
 } from "@/utils/cashu/wallet-recovery";
 import { persistBuyerP2pkEscrowRecord } from "@/utils/cashu/p2pk-escrow-records";
+import { generateKeys } from "@/utils/nostr/key-utilities";
 import {
   constructGiftWrappedEvent,
   constructMessageSeal,
@@ -56,7 +57,6 @@ import {
   sendGiftWrappedMessageEvent,
   getLocalStorageData,
   publishProofEvent,
-  generateKeys,
   getSavedAddresses,
 } from "@/utils/nostr/nostr-helper-functions";
 import { LightningAddress } from "@getalby/lightning-tools";
