+++ components/settings/storefront/section-editor.tsx
@@ -267,9 +267,7 @@ export default function SectionEditor({
                     onChange={(e) =>
                       update({
                         productLayout: e.target.value as
-                          | "grid"
-                          | "list"
-                          | "featured",
+                          "grid" | "list" | "featured",
                       })
                     }
                   >
