+++ packages/shopstr-mcp/src/types.ts
@@ -4,11 +4,7 @@ export type NostrEvent = Event;
 export type NostrFilter = Filter;
 
 export type ShippingOptionsType =
-  | "N/A"
-  | "Free"
-  | "Pickup"
-  | "Free/Pickup"
-  | "Added Cost";
+  "N/A" | "Free" | "Pickup" | "Free/Pickup" | "Added Cost";
 
 export type PricingBlock = {
   amount: number;
