+++ pages/api/cashu/validate-mint.ts
@@ -169,8 +169,7 @@ function hasBrowserCors(response: Response, requestOrigin: string): boolean {
 }
 
 type MintFetchResult =
-  | { ok: true; json: unknown }
-  | { ok: false; reason: "cors" | "request" };
+  { ok: true; json: unknown } | { ok: false; reason: "cors" | "request" };
 
 async function fetchJson(
   endpoint: string,
